Share the distance SQL the two canonical queries must agree on
NearestSchoolQuery and SchoolDistancesQuery now carry the same twenty-odd
lines of SELECT twice. They differ only in how many rows they keep per
course; the distance maths, the coordinate guard and the school columns are
identical, and they have to stay that way.

That is the reason to extract rather than tolerate the duplication. The two
feed the same page: the course page shows the nearest school's distance and
the ?debug panel lists every school's. If someone changed the divisor or
dropped the trailing `false` in one and not the other, the two would
quietly disagree about how far away the same school is - no test failure,
just numbers that no longer add up when a reviewer compares them.

Each query keeps its own DISTINCT ON, which is the part that actually
differs.
