+++ app/services/courses/nearest_school_query.rb
@@ -4,6 +4,8 @@ module Courses
   # Finds only nearest school per course
   #
   class NearestSchoolQuery
+    include CanonicalSchoolDistance
+
     def initialize(courses:, latitude:, longitude:)
       @courses = courses
       @latitude = latitude
@@ -37,8 +39,8 @@ module Courses
       Course
         .joins(schools: %i[gias_school provider_school])
         .where(id: @courses.map(&:id))
-        .where("gias_school.geo_location IS NOT NULL")
-        .select(schools_select_sql)
+        .where(GEOCODED_SCHOOL)
+        .select("DISTINCT ON (course.id) #{school_columns_sql}")
         .order("course.id, distance_to_search_location ASC, gias_school.id ASC")
     end
 
@@ -50,35 +52,6 @@ module Courses
         .where("site.longitude IS NOT NULL AND site.latitude IS NOT NULL")
         .select(select_sql)
         .order("course.id, distance_to_search_location ASC")
-    end
-
-    # geo_location is a stored generated column, so it is NULL whenever either
-    # coordinate is missing - one guard covers both. ST_Distance takes the
-    # trailing `false` to force sphere maths, matching the legacy
-    # ST_DistanceSphere path and Courses::Query#schools_location_scope, and
-    # metres are converted with the same 1609.344 the results page uses.
-    #
-    # The CASE reproduces Provider::School#location_name in SQL.
-    def schools_select_sql
-      <<~SQL.squish
-        DISTINCT ON (course.id) course.id AS course_id,
-        course.*,
-        provider_school.uuid AS provider_school_uuid,
-        CASE
-          WHEN provider_school.site_code = #{Course.connection.quote(Provider::School::MAIN_SITE_CODE)}
-          THEN gias_school.name || ' (Main Site)'
-          ELSE gias_school.name
-        END AS location_name,
-        gias_school.latitude,
-        gias_school.longitude,
-        ST_Distance(gias_school.geo_location, #{search_point}, false) / 1609.344 AS distance_to_search_location
-      SQL
-    end
-
-    def search_point
-      Course.sanitize_sql_array(
-        ["ST_SetSRID(ST_MakePoint(?::float, ?::float), 4326)::geography", Float(@longitude), Float(@latitude)],
-      )
     end
 
     def select_sql
